ai_service/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
werkzeug==2.3.6
